+++ Examples/Map+Positioning/Sources/Config.swift
@@ -12,8 +12,11 @@ import WemapMapSDK
 import WemapPositioningSDKVPSARKit
 
 enum PreferencesKey: String {
-    // versions
-    case mapVersion,
+
+    case enableHapticFeedback,
+         useWheelchair,
+         // versions
+         mapVersion,
          mapLibreVersion,
          positioningVersion,
          // Core constants
@@ -22,6 +25,7 @@ enum PreferencesKey: String {
          userLocationProjectionOnGraphEnabled,
          // Map constants
          switchLevelsAutomaticallyOnUserMovements,
+         staleTimeoutMilliseconds,
          // VPS constants
          slowConnectionSeconds,
          // Global navigation options
@@ -30,9 +34,21 @@ enum PreferencesKey: String {
          navigationRecalculationTimeInterval
 }
 
+enum AppConstants {
+    static var enableHapticFeedback: Bool = true
+    static var useWheelchair: Bool = false
+}
+
 func customKeysAndValues() -> [String: Any] {
-    
-    // Core
+
+    AppConstants.enableHapticFeedback = UserDefaults
+        .bool(forKey: .enableHapticFeedback, defaultValue: AppConstants.enableHapticFeedback)
+
+    AppConstants.useWheelchair = UserDefaults
+        .bool(forKey: .useWheelchair, defaultValue: AppConstants.useWheelchair)
+
+    // MARK: - Core
+
     CoreConstants.itineraryRecalculationEnabled = UserDefaults
         .bool(forKey: .itineraryRecalculationEnabled, defaultValue: CoreConstants.itineraryRecalculationEnabled)
     
@@ -42,15 +58,21 @@ func customKeysAndValues() -> [String: Any] {
     CoreConstants.userLocationProjectionOnGraphEnabled = UserDefaults
         .bool(forKey: .userLocationProjectionOnGraphEnabled, defaultValue: CoreConstants.userLocationProjectionOnGraphEnabled)
     
-    // Map
+    // MARK: - Map
+
     MapConstants.switchLevelsAutomaticallyOnUserMovements = UserDefaults
         .bool(forKey: .switchLevelsAutomaticallyOnUserMovements, defaultValue: MapConstants.switchLevelsAutomaticallyOnUserMovements)
     
-    // VPS
+    MapConstants.staleTimeoutMilliseconds = UserDefaults
+        .int(forKey: .staleTimeoutMilliseconds, defaultValue: MapConstants.staleTimeoutMilliseconds)
+
+    // MARK: - VPS
+
     VPSARKitConstants.slowConnectionSeconds = UserDefaults
         .int(forKey: .slowConnectionSeconds, defaultValue: VPSARKitConstants.slowConnectionSeconds)
 
-    // Versions
+    // MARK: - Versions
+
     let specificKeysAndValues: [PreferencesKey: Any] = [
         .mapVersion: Bundle.map.version,
         .mapLibreVersion: Bundle.mapLibre.version,