+++ Examples/Positioning+AR/Sources/Config.swift
@@ -11,11 +11,6 @@ import WemapCoreSDK
 import WemapGeoARSDK
 import WemapPositioningSDKVPSARKit
 
-enum ARGlobals {
-    // Saint Roch
-    static let origin = Coordinate(latitude: 43.60484182694762, longitude: 3.880688039628751, level: 0) // entrance
-}
-
 enum PreferencesKey: String {
     case positioningVersion,
          geoARVersion
@@ -28,6 +23,5 @@ func customKeysAndValues() -> [String: Any] {
         .geoARVersion: Bundle.geoAR.version
     ]
     
-    let dict = Dictionary(uniqueKeysWithValues: specificKeysAndValues.map { ($0.rawValue, $1) })
-    return dict
+    return Dictionary(uniqueKeysWithValues: specificKeysAndValues.map { ($0.rawValue, $1) })
 }